Rename var files to config files. Add template config structure
