+++ config/testdata/var2.hcl
